refactor: CouponService 로 통합
